pyproject.toml
[tool.poetry]
name = "cqrs-pattern"
version = "0.1.0"
description = ""
authors = ["harut <har.avetisyan2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
aio-pika = "^9.4.1"
prometheus-fastapi-instrumentator = "^7.0.0"
uvloop = "^0.19.0"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
